+++ src/algebra/expression.h
@@ -41,7 +41,7 @@ struct Expression {
   size_t size() const;
 
   Expression& truncate_by_size(size_t max_size);
-  Expression& truncate_by_norm(float min_norm);
+  Expression& truncate_by_norm(double min_norm);
 
   Expression& filter_by_size(size_t size);
